split google auth routes and services out into their own files, reworked jwt into their own services, added additional logs and comments, updated openapi docs
