+++ pkg/types/context.go
@@ -1,0 +1,4 @@
+package types
+
+// ContextKey alias type from string, to be passed via http.Request contexts.
+type ContextKey string
