+++ pkg/types/roles.go
@@ -1,5 +1,6 @@
 package types
 
+// Role representing the permission level of a user within the system.
 type Role string
 
 const (
